fix(item-batches): /expired 500 on every call -- function read products.name (column is nama_produk); sen kept
Found by the bigint-class sweep (23 Sep): get_expired_batches() -> "column i.name does not exist" on
every call, so GET /api/item-batches/expired answered 500 always. FE-reachable: the BatchExpiry screen
(ChatPanel 'batch_kedaluwarsa'). Columns checked in information_schema (products.nama_produk).
Also total_value was (quantity * unit_cost)::BIGINT -- money cast to integer.

V297: DROP + CREATE get_expired_batches: i.nama_produk (COALESCEd, LEFT JOIN vs required str),
total_value NUMERIC(18,2) = ROUND(qty * unit_cost, 2). Schemas ExpiredBatch / ExpiredBatchesResponse
total_value int -> float; router sums Decimal -> float (was int()).
DEPLOY: code FIRST, then V297 (old function 500s regardless; the new function with the old int schema
500s on any value with sen -- measured, sab2).

Gate (scratch; fixture: warehouse + product + batch expired yesterday, unit_cost 12345.67, stock 3):
  base RED  : 500 (column i.name)
  sab RED   : V297 with the OLD column name i.name (numeric fixed) -> 500
  sab2 RED  : real V297 + OLD code -> 500, pydantic int_from_float on 37037.01 (traceback read)
  new GREEN : 200, product name, total_value 37037.01

+++ backend/api_gateway/app/schemas/item_batches.py
@@ -212,7 +212,7 @@ class ExpiredBatch(BaseModel):
     expiry_date: date
     days_expired: int
     quantity: Decimal
-    total_value: int
+    total_value: float  # V297: NUMERIC(18,2); int membuang sen
     warehouse_id: UUID
     warehouse_name: str
 
@@ -222,7 +222,7 @@ class ExpiredBatchesResponse(BaseModel):
     success: bool = True
     data: List[ExpiredBatch]
     total: int
-    total_value: int
+    total_value: float
 
 
 # ============================================================================
